requirements.txt
# Exchanger.py - Universal Integration Platform
# Объединенные зависимости для всех компонентов

# Core dependencies для всех сервисов
pika==1.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
requests==2.32.4
loguru==0.7.3
python-dotenv==1.1.1

# Universal Worker specific
camunda-external-task-client-python3==4.5.0

# Task Creator specific
# (все зависимости уже включены в core)

# Task Tracker specific (планируется)
# schedule==1.2.0
# croniter==1.4.1

# Camunda-StormBPMN Sync Module dependencies
lxml>=4.9.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
prometheus-client>=0.17.0
python-dateutil>=2.8.0
tenacity>=8.2.0

# Development dependencies (опционально)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# flake8==6.1.0 